+++ CHANGELOG.md
@@ -2,3 +2,7 @@ Fixed:
 - Checking for update every 10 Minutes not working
 - Background album art is cut off (#116)
 - Tall context menus not scrollable (#114)
+- Default folder images not working
+
+Improved:
+- Moved `Hide Ads` into `Main Settings`